codex: Review - restore projects stack depth
